% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exceptions.R
\name{abort_redivis_error}
\alias{abort_redivis_error}
\title{Abort with a Redivis error}
\usage{
abort_redivis_error(
  message,
  class = "redivis_error",
  ...,
  call = rlang::caller_env()
)
}
\description{
Abort with a Redivis error
}
\keyword{internal}
