% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{authenticate}
\alias{authenticate}
\title{authenticate}
\usage{
authenticate(scope = NULL, force_reauthentication = FALSE)
}
\value{
void
}
\description{
Manually authenticate the current session with your Redivis credentials. Authentication normally happens automatically, and this method does not need to be called directly in most use cases.
}
\examples{
redivis::authenticate(force_reauthentication=FALSE)
}
