% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{file}
\alias{file}
\title{file}
\usage{
file(id)
}
\arguments{
\item{id}{The id of the file}
}
\value{
class<File>
}
\description{
Reference a file stored on Redivis
}
\examples{
file <- redivis::file("s335-8ey8zt7bx.qKmzpdttY2ZcaLB0wbRB7A")$download()
}
