% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{organization}
\alias{organization}
\title{organization}
\usage{
organization(name)
}
\arguments{
\item{name}{The organization's username}
}
\value{
class<user>
}
\description{
DEPRECATED: please use redivis$organization
}
\examples{
redivis::organization('demo_organization')$dataset('some_dataset')$table('a_table')$to_tibble(max_results=100)

We can also construct a query scoped to a particular dataset, removing the need to fully qualify table names
redivis::user('my_username')$workflow('my_workflow')$query("SELECT * FROM table_1 INNER JOIN table_2 ON id")$to_tibble()
}
