% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{query}
\alias{query}
\title{query}
\usage{
query(query = "", default_workflow = NULL, default_dataset = NULL)
}
\arguments{
\item{query}{The query string to execute}
}
\value{
class<Query>
}
\description{
Create a Redivis SQL query
}
\examples{
output_table <- redivis::query(query = 'SELECT 1 + 1 AS two')$to_tibble()
}
