% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{table}
\alias{table}
\title{table}
\usage{
table(name)
}
\arguments{
\item{name}{The table's username}
}
\value{
class<table>
}
\description{
DEPRECATED: please use redivis$table
}
\examples{
redivis::table('a_table')$to_tibble(max_results=100)

}
