% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{user}
\alias{user}
\title{user}
\usage{
user(name)
}
\arguments{
\item{name}{The user's username}
}
\value{
class<user>
}
\description{
Reference user-owned resources on Redivis
}
\examples{
redivis::user('my_username')$workflow('my_workflow')$table('my_table')$to_tibble(max_results=100)

We can also construct a query scoped to a particular workflow, removing the need to fully qualify table names
redivis::user('my_username')$workflow('my_workflow')$query("SELECT * FROM table_1 INNER JOIN table_2 ON id")$to_tibble()
}
