% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{current_notebook}
\alias{current_notebook}
\title{current_notebook}
\usage{
current_notebook()
}
\value{
class<Notebook>
}
\description{
A reference to the current Redivis notebook. Will be NULL if not running in a Redivis notebook environment.
}
\examples{
redivis::current_notebook()$create_output_table(df)
}
