% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{organization}
\alias{organization}
\title{organization}
\usage{
organization(name)
}
\arguments{
\item{name}{The organization's username}
}
\value{
class<user>
}
\description{
Reference organization-owned resources on Redivis
}
\examples{
redivis::organization('demo_organization')$dataset('some_dataset')$table('a_table')$to_dataframe(limit=100)

We can also construct a query scoped to a particular dataset, removing the need to fully qualify table names
redivis::user('my_username')$project('my_project')$query("SELECT * FROM table_1 INNER JOIN table_2 ON id")
}
