% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{query}
\alias{query}
\title{query}
\usage{
query(query = "")
}
\arguments{
\item{query}{The query string to execute}
}
\value{
A data frame object containing the query results
}
\description{
Execute a redivis query
}
\examples{
output_table <- query(query = 'SELECT 1 + 1 AS two')
}
