% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\docType{data}
\name{redivis}
\alias{redivis}
\title{redivis}
\format{
An object of class \code{list} of length 8.
}
\usage{
redivis
}
\value{
list
}
\description{
Redivis wrapper, all primary methods accessible via $
}
\examples{
dataset <- redivis$user("username")$dataset("dataset_name")
}
\keyword{datasets}
